+++ src/main/java/com/example/ProyectoBancoJPA/model/Transaccion.java
@@ -11,7 +11,7 @@ public class Transaccion {
     @Id
     @GeneratedValue(strategy = GenerationType.SEQUENCE )
     @SequenceGenerator(name = "transaccion_generator",allocationSize = 1 )
-    private long idtransaccion;
+    private Integer idtransaccion;
 
     @Column(name = "fechaTransaccion",nullable = false)
     private Date fechaTransaccion;
@@ -25,13 +25,13 @@ public class Transaccion {
     public Transaccion() {
     }
 
-    public Transaccion(long idTransacion, Date fechaTransaccion, Double monto,String tipo){
+    public Transaccion(Integer idTransacion, Date fechaTransaccion, Double monto,String tipo){
         this.fechaTransaccion = fechaTransaccion;
         this.monto = monto;
         this.tipo = tipo;
     }
 
-    public long getIdtransaccion() {
+    public Integer getIdtransaccion() {
         return idtransaccion;
     }
